+++ components/core/di/src/main/java/com/flipperdevices/core/di/ApplicationParams.kt
@@ -4,5 +4,6 @@ import android.app.Activity
 import kotlin.reflect.KClass
 
 data class ApplicationParams(
-    val startApplicationClass: KClass<out Activity>
+    val startApplicationClass: KClass<out Activity>,
+    val version: String
 )