+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/composable/ComposableSettings.kt
@@ -1,6 +1,7 @@
 package com.flipperdevices.settings.impl.composable
 
 import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
@@ -10,6 +11,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import androidx.navigation.NavHostController
@@ -17,11 +19,11 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.flipperdevices.core.ui.R as DesignSystem
 import com.flipperdevices.settings.impl.R
+import com.flipperdevices.settings.impl.composable.category.BugReportCategory
 import com.flipperdevices.settings.impl.composable.category.DebugCategory
 import com.flipperdevices.settings.impl.composable.category.ExperimentalCategory
-import com.flipperdevices.settings.impl.composable.category.GeneralCategory
-import com.flipperdevices.settings.impl.composable.elements.Category
-import com.flipperdevices.settings.impl.composable.elements.SimpleElement
+import com.flipperdevices.settings.impl.composable.category.VersionCategory
+import com.flipperdevices.settings.impl.composable.elements.AppBar
 import com.flipperdevices.settings.impl.model.NavGraphRoute
 import com.flipperdevices.settings.impl.viewmodels.SettingsViewModel
 
@@ -48,35 +50,19 @@ fun ComposableCommonSetting(
     navController: NavController,
     settingsViewModel: SettingsViewModel
 ) {
+    val context = LocalContext.current
     val settings by settingsViewModel.getState().collectAsState()
 
     Column(
-        Modifier
+        modifier = Modifier
             .verticalScroll(rememberScrollState())
-            .background(colorResource(DesignSystem.color.background))
+            .background(colorResource(DesignSystem.color.background)),
+        verticalArrangement = Arrangement.spacedBy(14.dp)
     ) {
-        Category(R.string.category_general)
-
-        GeneralCategory(
-            settings = settings,
-            onSwitchExperimental = settingsViewModel::onSwitchExperimental,
-            onSwitchDebug = settingsViewModel::onSwitchDebug
-        )
-
-        if (settings.enabledExperimentalFunctions) {
-            Category(titleId = R.string.category_experimental)
-            ExperimentalCategory(settings, navController)
-        }
-
-        if (settings.enabledDebugSettings) {
-            Category(titleId = R.string.category_debug)
-            DebugCategory(settings, navController)
-        }
-
-        val context = LocalContext.current
-        SimpleElement(
-            titleId = R.string.debug_shake2report_open,
-            onClick = { settingsViewModel.onReportBug(context) }
-        )
+        AppBar(R.string.options)
+        DebugCategory(settings, navController, settingsViewModel)
+        ExperimentalCategory(settings, navController, settingsViewModel)
+        BugReportCategory(onClick = { settingsViewModel.onReportBug(context) })
+        VersionCategory(version = settingsViewModel.versionApp())
     }
 }