+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/composable/elements/SimpleElement.kt
@@ -1,18 +1,16 @@
 package com.flipperdevices.settings.impl.composable.elements
 
-import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.Icon
-import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.runtime.Composable
@@ -20,37 +18,29 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import com.flipperdevices.core.ui.R as DesignSystem
+import com.flipperdevices.settings.impl.R
 
 @Composable
 fun SimpleElement(
     modifier: Modifier = Modifier,
-    @DrawableRes iconId: Int? = null,
     @StringRes titleId: Int? = null,
     @StringRes descriptionId: Int? = null,
-    onClick: (() -> Unit)? = null
-) {
-    SimpleElement(
-        modifier,
-        iconId,
-        titleId?.let { stringResource(it) },
-        descriptionId?.let { stringResource(it) },
-        onClick
+    onClick: (() -> Unit)? = null,
+    titleTextStyle: TextStyle = TextStyle(
+        fontWeight = FontWeight.W400,
+        fontSize = 14.sp
     )
-}
-
-@Composable
-fun SimpleElement(
-    modifier: Modifier = Modifier,
-    @DrawableRes iconId: Int? = null,
-    title: String? = null,
-    description: String? = null,
-    onClick: (() -> Unit)? = null
 ) {
+    val title: String? = titleId?.let { stringResource(id = it) }
+    val description = descriptionId?.let { stringResource(id = it) }
+
     var rowModifier = modifier
         .heightIn(min = 48.dp)
         .padding(all = 12.dp)
@@ -68,7 +58,6 @@ fun SimpleElement(
         rowModifier,
         verticalAlignment = Alignment.CenterVertically
     ) {
-        ElementIcon(iconId)
         Column(
             modifier = Modifier,
             verticalArrangement = Arrangement.Center
@@ -76,30 +65,19 @@ fun SimpleElement(
             if (title != null) {
                 Text(
                     text = title,
-                    style = MaterialTheme.typography.subtitle1
+                    style = titleTextStyle
                 )
             }
+            Spacer(modifier = Modifier.height(1.dp))
             if (description != null) {
                 Text(
                     text = description,
-                    style = MaterialTheme.typography.body2,
+                    fontSize = 12.sp,
                     color = colorResource(DesignSystem.color.black_40)
                 )
             }
         }
     }
-}
-
-@Composable
-private fun ElementIcon(@DrawableRes iconId: Int?) {
-    if (iconId == null) {
-        return
-    }
-    Icon(
-        modifier = Modifier.size(size = 42.dp),
-        painter = painterResource(iconId),
-        contentDescription = null
-    )
 }
 
 @Preview(
@@ -109,5 +87,9 @@ private fun ElementIcon(@DrawableRes iconId: Int?) {
 @Composable
 @Suppress("UnusedPrivateMember")
 private fun SimpleElementPreview() {
-    SimpleElement(Modifier, null, "Title", "Description", {})
+    SimpleElement(
+        modifier = Modifier,
+        titleId = R.string.experimental_screen_streaming,
+        descriptionId = R.string.experimental_screen_streaming_desc
+    )
 }