+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/fragments/SettingsFragment.kt
@@ -19,7 +19,7 @@ class SettingsFragment : ComposeFragment(), OnBackPressListener {
         ComposableSettings(navController = navController)
     }
 
-    override fun getStatusBarColor(): Int = DesignSystem.color.background
+    override fun getStatusBarColor(): Int = DesignSystem.color.accent
 
     override fun onBackPressed(): Boolean {
         val currentDestination = navController.currentDestination ?: return false